[FIXED JENKINS-18165]
AbstractProject->GitHubRepositoryName[] conversion is useful outside trigger

As indicated by the pull request #27, it is useful to be able to figure
out a list of repositories associated with the project even in the
absence of the trigger setting.

So I moved out this code into GitHubRepositoryName. This eliminated the
need for the GitHubCommitNotifier to require a GitHubPushTrigger, which
is the motivation for the pull request #27.