Let GithubProjectProperty be applied to any Job.
Normalizing databinding.
The projectUrl property was defined in the constructor to be a String, but in the getter to be a GithubUrl,
which made no sense and breaks tools looking to work with describables without GUI forms.
Changing the getter type would be the simplest fix, but potentially incompatible.
Instead changing the property name so that the existing getter can be left alone.