don't be platform specific in checkstyle.
While notable the existing checkstyle means the build is DOA on windows
that uses CRLF as a system linefeed.

Whilst this change enforces unix termination, it will break windows
clients configured with `core.autocrlf=true" but changing files can lead
to platform specific test failures (especially with resource files under
test) so EOL conversion is in this persons eyes evil and to be avoided at
all costs.

It would probably be worthwile to setup .gitattributes in the root of the
repository to stop any eol conversion - but I shall leave that up to the
owners for an excersize at their leisure.