[JENKINS-36445] Reduce length of displayed SHA1 in changes
Most git repository browsing systems (bitbucket, cgit, github, gitweb)
present a short form of the SHA1 (commonly the first 7 characters) with
the clickable hyperlink that will take them to a page that includes the
full SHA1 of the commit.

This change reduces the visible text to the first 7 characters of the
SHA1, consistent with those other repository browsers.  It also adds
assertions that format is used.