[JENKINS-36445] Show first 7 characters of SHA1 in change log hyperlink
Most git repository browsing systems (bitbucket, cgit, github, gitweb)
display a short form of the SHA1 (commonly the first 7 characters) with
a clickable hyperlink that will take them to a page that includes the
full SHA1.

This change reduces the visible SHA1 text in the list of changes to the
first 7 characters of the SHA1, consistent with those other repository
browsers. It also adds assertions that the expected format is used and
simplifies the existing annotator tests.