Print warning when repositories differ by case
There are many instances where people have run into hard to debug problems because Github webhooks in the plugin are case sensitive where the
rest of Jenkins is case insensitive. Ideally, I'd like to change the equality behavior to be case insensitive, but I recognize that could be backwards incompatible for
some people.

As a compromise, I'd like to add a WARN when comparing 2 repositories that would be equal but aren't because the case is different. At least in this case,
when users look at the logs, it will be obvious what is happening.