[JENKINS-53149] Calculate build data from downstream builds
Those downstream builds could contain, in example, a shared library which is loaded first in a pipeline. For that reason, we want to compare all remote URLs for each build data, with the real project name, to determine the proper build data. This way, the SHA returned in the build data will relate to the real project, and not from the first build data, which could be different to the real project.