Problem: [JENKINS-60901] GitHub manages hooks even when it has not been configured to do it (and complains a lot in the log)
Solution: Check if configured to manage hooks before trying to
register or unregister them (which in case of no credential set
up, results in a meaningful exception message plus a lengthy
but somewhat useless stacktrace). If nothing else, this reduces
I/O and storage overhead due to hook not configured and disabled.

Signed-off-by: Jim Klimov <jimklimov@gmail.com>