Fix to enable authentication with used HTTP proxy
Currently the Github plugin can not reach the Github server if the proxy used
by Jenkins requires authentication. Reason for that lies in the missing
configuration of proxy authenticator for the Github plugin.

This commit should resolve that issue by introducing a new HTTP client which is
preconfigured to handle the usage of Jenkins proxy configurations

This commit should resolve issue JENKINS-67806