fix: don't trigger poll unless target ref matches
The existing webhook handling code would trigger the GitSCM poll for the
default branch regardless of whether or not the target ref of the
webhook push event matched the branch specifier of the Job definition.
This is problematic and often leads to duplicate builds for the same
branch+revision if pushes to other branches happened to occur whilst
build(s) were already in progress for that branch. This is because the
GitSCM trigger just compares the HEAD revision of any branch matching
the branch specifier against the last *completed* build of that branch
(ignoring in-progress builds) and schedules a new build for it.

The code for this PR is mostly a rebased version of the original changes
proposed by @aserrallerios under
https://github.com/jenkinsci/github-plugin/pull/44 with some small
refinements.

Co-authored-by: Albert Serrallé Ríos <aserrallerios@gmail.com>