Test Java 11 directly
Address the concern that Oleg expressed in
https://github.com/jenkinsci/github-plugin/pull/368#pullrequestreview-1726071737

We've detected no regressions with the transition to testing Java 11
byte code as generated by Java 17 and Java 21 compilers in the 200+
plugins that have made the transition.

We test with Java 11 on older lines (Jenkins 2.401.x and Jenkins
2.414.x) in the plugin bill of materials while testing with Java 21 on
the most recent weekly release.  We've detected no issues with any of
the plugins that are included in the plugin BOM.

We want to reduce Jenkins infrastructure costs by not testing
configurations that are unlikely to detect issues.  Testing with Java
11 is unlikely to detect issues and adds infrastructure cost.

I've added Java 11 on Linux as a test configuration. I do not intend
to make that same addition on the other 100+ repositories where we're
testing with Java 21 and Java 17.  When a plugin maintainer asks to
continue testing Java 11, I'm willing to adapt that pull request.