Test with Java 21 (#368)
* Test with Java 21

Java 21 was released Sep 19, 2023. We want to announce full support
for Java 21 in early October and would like the most used plugins to
be compiled and tested with Java 21.

The acceptance test harness and plugin bill of materials tests are
already passing with Java 21. This is a further step to improve plugin
readiness for use with Java 21 and for development with Java 21.

The change intentionally tests only two Java configurations, Java 17
and Java 21 because we believe that the risk of a regression that only
affects Java 11 is low. We generate Java 11 byte code with the Java 17
and the Java 21 builds, so we're already testing Java 11 byte code.

Also updates parent pom to most current release and removes unneeded
hamcrest dependency declaration.

Supersedes pull request:

* #367
* #366
* #361
* #360
* #359
* #353
* #352

* Test Java 11 directly

Address the concern that Oleg expressed in
https://github.com/jenkinsci/github-plugin/pull/368#pullrequestreview-1726071737

We've detected no regressions with the transition to testing Java 11
byte code as generated by Java 17 and Java 21 compilers in the 200+
plugins that have made the transition.

We test with Java 11 on older lines (Jenkins 2.401.x and Jenkins
2.414.x) in the plugin bill of materials while testing with Java 21 on
the most recent weekly release.  We've detected no issues with any of
the plugins that are included in the plugin BOM.

We want to reduce Jenkins infrastructure costs by not testing
configurations that are unlikely to detect issues.  Testing with Java
11 is unlikely to detect issues and adds infrastructure cost.

I've added Java 11 on Linux as a test configuration. I do not intend
to make that same addition on the other 100+ repositories where we're
testing with Java 21 and Java 17.  When a plugin maintainer asks to
continue testing Java 11, I'm willing to adapt that pull request.

* Revert "Test Java 11 directly"

This reverts commit 6d8ef2ff8f3e37f3540c4274d320e624fc03d2d2.